+++ NEWS.md
@@ -1,6 +1,9 @@
 # cSEM 0.1.0:9000
 
-- `calculateAVE()` is now exported and supportes all of cSEM's `cSEMResults` classes.
+- `calculateGoF()` is now exported and supports all of cSEM's `cSEMResults` classes.
+  Breaking: the `.only_common_factors` argument has been removed.
+
+- `calculateAVE()` is now exported and supports all of cSEM's `cSEMResults` classes.
 
 - Bug fix: model-implied indicator correlation matrix for non-recursive models
   should now be calculated correctly (#264)